+++ packages/eslint-plugin-internal/src/rules/index.ts
@@ -1,7 +1,9 @@
 import noTypescriptDefaultImport from './no-typescript-default-import';
 import noTypescriptEstreeImport from './no-typescript-estree-import';
+import preferASTTypesEnum from './prefer-ast-types-enum';
 
 export default {
   'no-typescript-default-import': noTypescriptDefaultImport,
   'no-typescript-estree-import': noTypescriptEstreeImport,
+  'prefer-ast-types-enum': preferASTTypesEnum,
 };