+++ packages/eslint-plugin/src/rules/default-param-last.ts
@@ -12,6 +12,7 @@ export default createRule({
       description: 'Enforce default parameters to be last',
       category: 'Best Practices',
       recommended: false,
+      extendsBaseRule: true,
     },
     schema: [],
     messages: {