+++ packages/eslint-plugin/tests/rules/no-this-alias.test.ts
@@ -3,15 +3,15 @@ import rule from '../../src/rules/no-this-alias';
 import { RuleTester } from '../RuleTester';
 
 const idError = {
-  messageId: 'thisAssignment' as 'thisAssignment',
+  messageId: 'thisAssignment' as const,
   type: AST_NODE_TYPES.Identifier,
 };
 const destructureError = {
-  messageId: 'thisDestructure' as 'thisDestructure',
+  messageId: 'thisDestructure' as const,
   type: AST_NODE_TYPES.ObjectPattern,
 };
 const arrayDestructureError = {
-  messageId: 'thisDestructure' as 'thisDestructure',
+  messageId: 'thisDestructure' as const,
   type: AST_NODE_TYPES.ArrayPattern,
 };