+++ packages/experimental-utils/src/ts-eslint-scope/Referencer.ts
@@ -44,8 +44,8 @@ interface Referencer<SM extends ScopeManager> extends Visitor {
 
   AssignmentExpression(node: TSESTree.Node): void;
   CatchClause(node: TSESTree.Node): void;
-  Program(node: TSESTree.Node): void;
-  Identifier(node: TSESTree.Node): void;
+  Program(node: TSESTree.Program): void;
+  Identifier(node: TSESTree.Identifier): void;
   UpdateExpression(node: TSESTree.Node): void;
   MemberExpression(node: TSESTree.Node): void;
   Property(node: TSESTree.Node): void;